refactor(release): modularize monolithic release service
