+++ internal/services/release_service.go
@@ -3,12 +3,6 @@ package services
 import (
 	"context"
 	"fmt"
-	"os"
-	"path/filepath"
-	"regexp"
-	"strconv"
-	"strings"
-	"time"
 
 	"github.com/thomas-vilte/matecommit/internal/ai"
 	"github.com/thomas-vilte/matecommit/internal/config"
@@ -18,30 +12,8 @@ import (
 	"github.com/thomas-vilte/matecommit/internal/models"
 	"github.com/thomas-vilte/matecommit/internal/regex"
 	"github.com/thomas-vilte/matecommit/internal/vcs"
-	"golang.org/x/mod/semver"
 )
 
-const (
-	maxSearchDepth = 5
-	maxFilesToScan = 100
-)
-
-var ignoreDirs = map[string]bool{
-	".git":          true,
-	"node_modules":  true,
-	"vendor":        true,
-	"target":        true,
-	"dist":          true,
-	"build":         true,
-	".next":         true,
-	".nuxt":         true,
-	"__pycache__":   true,
-	".pytest_cache": true,
-	"venv":          true,
-	".venv":         true,
-}
-
-// releaseGitService defines only the methods needed by ReleaseService.
 type releaseGitService interface {
 	GetLastTag(ctx context.Context) (string, error)
 	GetCommitCount(ctx context.Context) (int, error)
@@ -59,7 +31,6 @@ type releaseGitService interface {
 	ValidateTagExists(ctx context.Context, tag string) error
 	GetRepoRoot(ctx context.Context) (string, error)
 }
-
 type ReleaseService struct {
 	git         releaseGitService
 	vcsClient   vcs.VCSClient
@@ -73,7 +44,6 @@ type ReleaseService struct {
 		lang    string
 	}
 }
-
 type ReleaseOption func(*ReleaseService)
 
 func WithReleaseVCSClient(vcs vcs.VCSClient) ReleaseOption {
@@ -93,7 +63,6 @@ func WithReleaseConfig(cfg *config.Config) ReleaseOption {
 		s.config = cfg
 	}
 }
-
 func NewReleaseService(
 	gitSvc releaseGitService,
 	opts ...ReleaseOption,
@@ -107,7 +76,6 @@ func NewReleaseService(
 	}
 	return s
 }
-
 func (s *ReleaseService) AnalyzeNextRelease(ctx context.Context) (*models.Release, error) {
 	log := logger.FromContext(ctx)
 
@@ -176,7 +144,6 @@ func (s *ReleaseService) AnalyzeNextRelease(ctx context.Context) (*models.Releas
 
 	return release, nil
 }
-
 func (s *ReleaseService) GenerateReleaseNotes(ctx context.Context, release *models.Release) (*models.ReleaseNotes, error) {
 	log := logger.FromContext(ctx)
 
@@ -201,7 +168,6 @@ func (s *ReleaseService) GenerateReleaseNotes(ctx context.Context, release *mode
 
 	return s.notesGen.GenerateNotes(ctx, release)
 }
-
 func (s *ReleaseService) PublishRelease(ctx context.Context, release *models.Release, notes *models.ReleaseNotes, draft bool, buildBinaries bool, progressCh chan<- models.BuildProgress) error {
 	log := logger.FromContext(ctx)
 
@@ -228,7 +194,6 @@ func (s *ReleaseService) PublishRelease(ctx context.Context, release *models.Rel
 
 	return nil
 }
-
 func (s *ReleaseService) CreateTag(ctx context.Context, version, message string) error {
 	log := logger.FromContext(ctx)
 
@@ -247,7 +212,6 @@ func (s *ReleaseService) CreateTag(ctx context.Context, version, message string)
 
 	return nil
 }
-
 func (s *ReleaseService) TagExists(ctx context.Context, version string) bool {
 	if s.git == nil {
 		return false
@@ -255,7 +219,6 @@ func (s *ReleaseService) TagExists(ctx context.Context, version string) bool {
 
 	return s.git.ValidateTagExists(ctx, version) == nil
 }
-
 func (s *ReleaseService) PushTag(ctx context.Context, version string) error {
 	log := logger.FromContext(ctx)
 
@@ -274,14 +237,12 @@ func (s *ReleaseService) PushTag(ctx context.Context, version string) error {
 
 	return nil
 }
-
 func (s *ReleaseService) GetRelease(ctx context.Context, version string) (*models.VCSRelease, error) {
 	if s.vcsClient == nil {
 		return nil, domainErrors.ErrVCSNotSupported
 	}
 	return s.vcsClient.GetRelease(ctx, version)
 }
-
 func (s *ReleaseService) UpdateRelease(ctx context.Context, version, body string) error {
 	log := logger.FromContext(ctx)
 
@@ -305,7 +266,6 @@ func (s *ReleaseService) UpdateRelease(ctx context.Context, version, body string
 
 	return nil
 }
-
 func (s *ReleaseService) EnrichReleaseContext(ctx context.Context, release *models.Release) error {
 	log := logger.FromContext(ctx)
 
@@ -354,1079 +314,6 @@ func (s *ReleaseService) EnrichReleaseContext(ctx context.Context, release *mode
 
 	return nil
 }
-
-func (s *ReleaseService) UpdateLocalChangelog(ctx context.Context, release *models.Release, notes *models.ReleaseNotes) error {
-	log := logger.FromContext(ctx)
-
-	repoRoot, err := s.git.GetRepoRoot(ctx)
-	if err != nil {
-		log.Error("failed to get repository root for changelog", "error", err)
-		return fmt.Errorf("failed to get repository root: %w", err)
-	}
-	changelogFile := filepath.Join(repoRoot, "CHANGELOG.md")
-
-	log.Debug("updating local changelog",
-		"version", release.Version,
-		"file", changelogFile)
-
-	if _, err := os.Stat(changelogFile); err == nil {
-		content, readErr := os.ReadFile(changelogFile)
-		if readErr == nil && strings.Contains(string(content), "## [Unreleased]") {
-			if err := s.MoveUnreleasedToVersion(changelogFile, release, notes); err != nil {
-				log.Warn("failed to move Unreleased section", "error", err)
-			} else {
-				log.Info("moved Unreleased section to new version", "version", release.Version)
-				if s.hasUnreleasedContent(content) {
-					return nil
-				}
-				log.Debug("unreleased section was empty, will add new version entry")
-			}
-		}
-	}
-
-	newContent := s.buildChangelogFromNotes(context.Background(), release, notes)
-
-	if err := s.prependToChangelog(changelogFile, newContent); err != nil {
-		log.Error("failed to update changelog",
-			"error", err,
-			"file", changelogFile)
-		return domainErrors.NewAppError(
-			domainErrors.TypeInternal,
-			"Failed to update CHANGELOG.md",
-			err,
-		).WithSuggestion(
-			"Make sure CHANGELOG.md is writable and not locked by another process.\n" +
-				"Try: chmod +w CHANGELOG.md",
-		)
-	}
-
-	if err := s.EnsureUnreleasedSection(changelogFile); err != nil {
-		log.Warn("failed to ensure Unreleased section", "error", err)
-	}
-
-	if warnings, err := s.ValidateChangelog(changelogFile); err == nil && len(warnings) > 0 {
-		log.Warn("CHANGELOG validation warnings detected", "count", len(warnings))
-		for _, warning := range warnings {
-			log.Warn("CHANGELOG warning", "type", warning.Type, "message", warning.Message)
-		}
-	}
-
-	log.Info("changelog updated successfully",
-		"file", changelogFile,
-		"version", release.Version)
-
-	return nil
-}
-
-func (s *ReleaseService) prependToChangelog(filename, newContent string) error {
-	content, err := os.ReadFile(filename)
-	if os.IsNotExist(err) {
-		header := "# Changelog\n\nAll notable changes to this project will be documented in this file.\n\n"
-		return os.WriteFile(filename, []byte(header+newContent), 0644)
-	}
-	if err != nil {
-		return domainErrors.NewAppError(
-			domainErrors.TypeInternal,
-			"Failed to read CHANGELOG.md",
-			err,
-		).WithSuggestion(
-			"Ensure CHANGELOG.md exists and is readable.\n" +
-				"Try: ls -la CHANGELOG.md",
-		)
-	}
-
-	current := string(content)
-
-	versionRegex := regexp.MustCompile(`## \[([^]]+)]`)
-	matches := versionRegex.FindStringSubmatch(newContent)
-	if len(matches) < 2 {
-		// If we can't extract version, fallback to old behavior
-		return s.prependToChangelogLegacy(filename, current, newContent)
-	}
-
-	newVersion := matches[1]
-
-	headerPattern := regexp.MustCompile(`(?m)^## \[` + regexp.QuoteMeta(newVersion) + `\].*$`)
-	if headerPattern.MatchString(current) {
-		log := logger.FromContext(context.Background())
-		log.Debug("version already exists in changelog, replacing",
-			"version", newVersion)
-
-		loc := headerPattern.FindStringIndex(current)
-		if loc != nil {
-			nextVersionPattern := regexp.MustCompile(`(?m)^## \[[^]]+]`)
-			nextLoc := nextVersionPattern.FindStringIndex(current[loc[1]:])
-
-			end := len(current)
-			if nextLoc != nil {
-				end = loc[1] + nextLoc[0]
-			}
-
-			before := current[:loc[0]]
-			after := current[end:]
-
-			before = strings.TrimRight(before, "\n")
-
-			var sb strings.Builder
-			sb.WriteString(before)
-			sb.WriteString("\n\n")
-			sb.WriteString(strings.TrimSpace(newContent))
-			sb.WriteString("\n\n")
-			sb.WriteString(strings.TrimLeft(after, "\n"))
-
-			result := sb.String()
-
-			result = s.consolidateLinkDefinitions(result)
-
-			return os.WriteFile(filename, []byte(result), 0644)
-		}
-	}
-
-	return s.prependToChangelogLegacy(filename, current, newContent)
-}
-
-// prependToChangelogLegacy is the original prepend logic
-func (s *ReleaseService) prependToChangelogLegacy(filename, current, newContent string) error {
-	var sb strings.Builder
-
-	versionPattern := regexp.MustCompile(`\n## \[[^]]+]`)
-	loc := versionPattern.FindStringIndex(current)
-
-	if loc != nil {
-		pre := current[:loc[0]]
-		post := current[loc[0]:]
-
-		sb.WriteString(strings.TrimSpace(pre))
-		sb.WriteString("\n\n")
-		sb.WriteString(strings.TrimSpace(newContent))
-		sb.WriteString("\n")
-		sb.WriteString(post)
-	} else {
-		if strings.HasPrefix(current, "# ") {
-			sb.WriteString(strings.TrimSpace(current))
-			sb.WriteString("\n\n")
-			sb.WriteString(strings.TrimSpace(newContent))
-			sb.WriteString("\n")
-		} else {
-			sb.WriteString(strings.TrimSpace(newContent))
-			sb.WriteString("\n\n")
-			sb.WriteString(strings.TrimSpace(current))
-		}
-	}
-
-	result := sb.String()
-
-	// Remove empty Unreleased sections that might be between versions
-	result = s.removeEmptyUnreleasedSections(result)
-
-	result = s.consolidateLinkDefinitions(result)
-
-	return os.WriteFile(filename, []byte(result), 0644)
-}
-
-// removeEmptyUnreleasedSections removes empty ## [Unreleased] sections that appear between versions
-func (s *ReleaseService) removeEmptyUnreleasedSections(content string) string {
-	emptyUnreleasedPattern := regexp.MustCompile(`(?s)## \[Unreleased\]\s*\n(## \[)`)
-	return emptyUnreleasedPattern.ReplaceAllString(content, "$1")
-}
-
-// consolidateLinkDefinitions removes duplicate link reference definitions
-func (s *ReleaseService) consolidateLinkDefinitions(content string) string {
-	linkDefPattern := regexp.MustCompile(`(?m)^\[([^]]+)]:\s*(.+)$`)
-
-	lines := strings.Split(content, "\n")
-	seen := make(map[string]string)
-	var result []string
-
-	for _, line := range lines {
-		matches := linkDefPattern.FindStringSubmatch(line)
-		if len(matches) == 3 {
-			version := matches[1]
-			url := matches[2]
-
-			if existingURL, exists := seen[version]; exists {
-				if existingURL != url {
-					log := logger.FromContext(context.Background())
-					log.Warn("duplicate link definition with different URL",
-						"version", version,
-						"existing_url", existingURL,
-						"new_url", url)
-				}
-				continue
-			}
-			seen[version] = url
-		}
-		result = append(result, line)
-	}
-
-	return strings.Join(result, "\n")
-}
-
-// EnsureUnreleasedSection ensures an Unreleased section exists in the CHANGELOG
-func (s *ReleaseService) EnsureUnreleasedSection(filename string) error {
-	content, err := os.ReadFile(filename)
-	if os.IsNotExist(err) {
-		header := `# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-`
-		return os.WriteFile(filename, []byte(header), 0644)
-	}
-	if err != nil {
-		return domainErrors.NewAppError(
-			domainErrors.TypeInternal,
-			"Failed to read CHANGELOG.md for Unreleased section",
-			err,
-		).WithSuggestion(
-			"Check if CHANGELOG.md is readable and not corrupted.\n" +
-				"Try: cat CHANGELOG.md",
-		)
-	}
-
-	current := string(content)
-
-	if strings.Contains(current, "## [Unreleased]") {
-		return nil
-	}
-
-	versionPattern := regexp.MustCompile(`\n## \[[^]]+]`)
-	loc := versionPattern.FindStringIndex(current)
-
-	if loc == nil {
-		current = strings.TrimSpace(current) + "\n\n## [Unreleased]\n\n"
-	} else {
-		pre := current[:loc[0]]
-		post := current[loc[0]:]
-		current = strings.TrimSpace(pre) + "\n\n## [Unreleased]\n\n" + post
-	}
-
-	return os.WriteFile(filename, []byte(current), 0644)
-}
-
-// parseUnreleasedSection extracts the Unreleased section content
-func (s *ReleaseService) parseUnreleasedSection(content string) string {
-	unreleasedPattern := regexp.MustCompile(`(?s)## \[Unreleased]\s*\n(.*?)(?:## \[|$)`)
-	matches := unreleasedPattern.FindStringSubmatch(content)
-
-	if len(matches) < 2 {
-		return ""
-	}
-
-	return strings.TrimSpace(matches[1])
-}
-
-// hasUnreleasedContent checks if the Unreleased section has actual content
-func (s *ReleaseService) hasUnreleasedContent(content []byte) bool {
-	return s.parseUnreleasedSection(string(content)) != ""
-}
-
-// MoveUnreleasedToVersion moves Unreleased section content to a new version
-func (s *ReleaseService) MoveUnreleasedToVersion(filename string, release *models.Release, notes *models.ReleaseNotes) error {
-	content, err := os.ReadFile(filename)
-	if err != nil {
-		return domainErrors.NewAppError(
-			domainErrors.TypeInternal,
-			"Failed to read CHANGELOG.md for Unreleased migration",
-			err,
-		).WithSuggestion(
-			"Ensure CHANGELOG.md exists and is readable.",
-		)
-	}
-
-	current := string(content)
-
-	unreleasedContent := s.parseUnreleasedSection(current)
-
-	log := logger.FromContext(context.Background())
-	log.Debug("parsed unreleased section",
-		"unreleased_content", unreleasedContent,
-		"unreleased_content_length", len(unreleasedContent),
-		"is_empty", unreleasedContent == "")
-
-	if unreleasedContent == "" {
-		log.Info("unreleased section is empty, skipping migration")
-		return nil
-	}
-
-	log.Info("moving Unreleased section to new version",
-		"version", release.Version,
-		"unreleased_content_length", len(unreleasedContent))
-
-	versionEntry := s.buildChangelogFromNotes(context.Background(), release, notes)
-
-	versionEntry = strings.TrimSpace(versionEntry) + "\n\n" + unreleasedContent + "\n"
-
-	unreleasedPattern := regexp.MustCompile(`(?s)## \[Unreleased]\s*\n.*?(\n## \[|$)`)
-	current = unreleasedPattern.ReplaceAllString(current, "$1")
-
-	log.Debug("writing modified changelog without unreleased section",
-		"filename", filename,
-		"content_length", len(current))
-
-	if err := os.WriteFile(filename, []byte(current), 0644); err != nil {
-		return domainErrors.NewAppError(
-			domainErrors.TypeInternal,
-			"Failed to write CHANGELOG.md during Unreleased migration",
-			err,
-		).WithSuggestion(
-			"Check file permissions and disk space.\n" +
-				"Try: df -h . && chmod +w CHANGELOG.md",
-		)
-	}
-
-	log.Debug("changelog written successfully, prepending new version")
-
-	if err := s.prependToChangelog(filename, versionEntry); err != nil {
-		return domainErrors.NewAppError(
-			domainErrors.TypeInternal,
-			"Failed to prepend new version during Unreleased migration",
-			err,
-		).WithSuggestion(
-			"The Unreleased section was removed but the new version couldn't be added.\n" +
-				"You may need to manually restore CHANGELOG.md from git.",
-		)
-	}
-
-	return s.EnsureUnreleasedSection(filename)
-}
-
-// ChangelogWarning represents a validation warning
-type ChangelogWarning struct {
-	Type    string
-	Message string
-}
-
-// validateChangelogEntry validates a CHANGELOG entry and returns warnings
-func (s *ReleaseService) validateChangelogEntry(content string, version string) []ChangelogWarning {
-	var warnings []ChangelogWarning
-
-	datePattern := regexp.MustCompile(`## \[` + regexp.QuoteMeta(version) + `\] - \d{4}-\d{2}-\d{2}`)
-	if !datePattern.MatchString(content) {
-		warnings = append(warnings, ChangelogWarning{
-			Type:    "missing_date",
-			Message: fmt.Sprintf("Version %s is missing a date or date is not in ISO 8601 format (YYYY-MM-DD)", version),
-		})
-	}
-
-	linkPattern := regexp.MustCompile(`\[` + regexp.QuoteMeta(version) + `\]:\s*https?://`)
-	if !linkPattern.MatchString(content) {
-		warnings = append(warnings, ChangelogWarning{
-			Type:    "missing_link",
-			Message: fmt.Sprintf("Version %s is missing a comparison link", version),
-		})
-	}
-
-	if !strings.Contains(content, "###") {
-		warnings = append(warnings, ChangelogWarning{
-			Type:    "no_sections",
-			Message: fmt.Sprintf("Version %s has no sections (###). Consider organizing changes into sections.", version),
-		})
-	}
-
-	versionHeaderPattern := regexp.MustCompile(`(?s)## \[` + regexp.QuoteMeta(version) + `\].*?\n\n(.*?)(?:## \[|$)`)
-	matches := versionHeaderPattern.FindStringSubmatch(content)
-	if len(matches) > 1 {
-		actualContent := strings.TrimSpace(matches[1])
-		actualContent = regexp.MustCompile(`(?m)^\[.*?]:.*$`).ReplaceAllString(actualContent, "")
-		actualContent = strings.TrimSpace(actualContent)
-
-		if len(actualContent) < 50 {
-			warnings = append(warnings, ChangelogWarning{
-				Type:    "short_content",
-				Message: fmt.Sprintf("Version %s has very little content. Consider adding more details.", version),
-			})
-		}
-	}
-
-	return warnings
-}
-
-// ValidateChangelog validates the entire CHANGELOG file
-func (s *ReleaseService) ValidateChangelog(filename string) ([]ChangelogWarning, error) {
-	content, err := os.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	var allWarnings []ChangelogWarning
-	current := string(content)
-
-	versionPattern := regexp.MustCompile(`## \[([^]]+)]`)
-	matches := versionPattern.FindAllStringSubmatch(current, -1)
-
-	for _, match := range matches {
-		if len(match) > 1 {
-			version := match[1]
-			if version == "Unreleased" {
-				continue
-			}
-
-			warnings := s.validateChangelogEntry(current, version)
-			allWarnings = append(allWarnings, warnings...)
-		}
-	}
-
-	return allWarnings, nil
-}
-
-func (s *ReleaseService) analyzeDependencyChanges(ctx context.Context, release *models.Release) ([]models.DependencyChange, error) {
-	if s.vcsClient == nil {
-		return []models.DependencyChange{}, nil
-	}
-	return s.depAnalyzer.AnalyzeAll(ctx, s.vcsClient, release.PreviousVersion, release.Version)
-}
-
-// categorizeCommits categorizes commits according to conventional commits
-func (s *ReleaseService) categorizeCommits(release *models.Release) {
-	for _, commit := range release.AllCommits {
-		msg := commit.Message
-		lines := strings.Split(msg, "\n")
-		firstLine := lines[0]
-
-		prMatch := regex.GitHubPR.FindStringSubmatch(firstLine)
-		prNumber := ""
-		if len(prMatch) > 1 {
-			prNumber = prMatch[1]
-		}
-
-		hasBreaking := false
-		for _, line := range lines[1:] {
-			if regex.BreakingChange.MatchString(line) {
-				hasBreaking = true
-				break
-			}
-		}
-
-		matches := regex.ConventionalCommit.FindStringSubmatch(firstLine)
-		if len(matches) > 0 {
-			commitType := matches[1]
-			scope := matches[3]
-			breaking := matches[4] == "!" || hasBreaking
-			description := strings.TrimSpace(matches[5])
-
-			item := models.ReleaseItem{
-				Type:        commitType,
-				Scope:       scope,
-				Description: description,
-				Breaking:    breaking,
-				PRNumber:    prNumber,
-			}
-
-			switch commitType {
-			case "feat":
-				if breaking {
-					release.Breaking = append(release.Breaking, item)
-				} else {
-					release.Features = append(release.Features, item)
-				}
-			case "fix":
-				release.BugFixes = append(release.BugFixes, item)
-			case "docs":
-				release.Documentation = append(release.Documentation, item)
-			case "perf", "refactor":
-				release.Improvements = append(release.Improvements, item)
-			default:
-				release.Other = append(release.Other, item)
-			}
-		} else {
-			item := models.ReleaseItem{
-				Type:        "other",
-				Description: firstLine,
-				PRNumber:    prNumber,
-			}
-			release.Other = append(release.Other, item)
-		}
-	}
-}
-
-// calculateVersion calculates the new version based on semantic versioning
-func (s *ReleaseService) calculateVersion(currentTag string, release *models.Release) (string, models.VersionBump) {
-	matches := regex.SemVer.FindStringSubmatch(currentTag)
-
-	major, minor, patch := 0, 0, 0
-	if len(matches) >= 4 {
-		major, _ = strconv.Atoi(matches[1])
-		minor, _ = strconv.Atoi(matches[2])
-		patch, _ = strconv.Atoi(matches[3])
-	}
-
-	bump := models.PatchBump
-
-	if len(release.Breaking) > 0 {
-		major++
-		minor = 0
-		patch = 0
-		bump = models.MajorBump
-	} else if len(release.Features) > 0 {
-		// MINOR: new features
-		minor++
-		patch = 0
-		bump = models.MinorBump
-	} else if len(release.BugFixes) > 0 || len(release.Improvements) > 0 {
-		// PATCH: bug fixes or improvements
-		patch++
-		bump = models.PatchBump
-	}
-
-	newVersion := fmt.Sprintf("v%d.%d.%d", major, minor, patch)
-	return newVersion, bump
-}
-
-func (s *ReleaseService) generateBasicNotes(release *models.Release) *models.ReleaseNotes {
-	title := fmt.Sprintf("Version %s", release.Version)
-
-	summary := fmt.Sprintf("This release includes %d new features, %d bug fixes",
-		len(release.Features), len(release.BugFixes))
-
-	if len(release.Breaking) > 0 {
-		summary += fmt.Sprintf(" and %d breaking changes", len(release.Breaking))
-	}
-
-	changelog := s.buildChangelog(release)
-
-	return &models.ReleaseNotes{
-		Title:       title,
-		Summary:     summary,
-		Changelog:   changelog,
-		Recommended: release.VersionBump,
-	}
-}
-
-// buildChangelogFromNotes formats the changelog using AI-generated highlights and semantic sections
-func (s *ReleaseService) buildChangelogFromNotes(ctx context.Context, release *models.Release, notes *models.ReleaseNotes) string {
-	var sb strings.Builder
-
-	tagDate, err := s.git.GetTagDate(ctx, release.Version)
-	if err != nil || tagDate == "" {
-		tagDate = time.Now().Format("2006-01-02")
-	}
-
-	owner, repo, provider, _ := s.git.GetRepoInfo(ctx)
-
-	versionHeader := fmt.Sprintf("## [%s] - %s", release.Version, tagDate)
-
-	if provider == "github" && owner != "" && repo != "" {
-		compareURL := ""
-		if release.PreviousVersion != "" {
-			compareURL = fmt.Sprintf("https://github.com/%s/%s/compare/%s...%s", owner, repo, release.PreviousVersion, release.Version)
-		} else {
-			compareURL = fmt.Sprintf("https://github.com/%s/%s/releases/tag/%s", owner, repo, release.Version)
-		}
-		versionHeader += fmt.Sprintf("\n\n[%s]: %s", release.Version, compareURL)
-	}
-
-	sb.WriteString(versionHeader + "\n\n")
-
-	if notes.Summary != "" {
-		sb.WriteString(fmt.Sprintf("%s\n\n", notes.Summary))
-	}
-
-	usedReferences := make(map[string]bool)
-
-	if len(notes.Sections) > 0 {
-		for _, section := range notes.Sections {
-			if section.Title != "" && len(section.Items) > 0 {
-				sb.WriteString(fmt.Sprintf("### %s\n\n", section.Title))
-				for _, item := range section.Items {
-					sb.WriteString(fmt.Sprintf("- %s\n", s.formatNoteBulletWithReference(release, item, owner, repo, provider, usedReferences)))
-				}
-				sb.WriteString("\n")
-			}
-		}
-	} else if len(notes.Highlights) > 0 {
-		sb.WriteString("### ✨ Highlights\n\n")
-		for _, highlight := range notes.Highlights {
-			sb.WriteString(fmt.Sprintf("- %s\n", s.formatNoteBulletWithReference(release, highlight, owner, repo, provider, usedReferences)))
-		}
-		sb.WriteString("\n")
-	}
-
-	if len(notes.BreakingChanges) > 0 {
-		sb.WriteString("### ⚠️ Breaking Changes\n\n")
-		for _, bc := range notes.BreakingChanges {
-			sb.WriteString(fmt.Sprintf("- %s\n", s.formatNoteBulletWithReference(release, bc, owner, repo, provider, usedReferences)))
-		}
-		sb.WriteString("\n")
-	}
-
-	references := s.collectReleaseReferences(release, owner, repo, provider, usedReferences)
-	if len(references) > 0 {
-		sb.WriteString("### References\n\n")
-		for _, reference := range references {
-			sb.WriteString(fmt.Sprintf("- %s\n", reference))
-		}
-		sb.WriteString("\n")
-	}
-
-	return sb.String()
-}
-
-func (s *ReleaseService) formatNoteBulletWithReference(release *models.Release, itemText, owner, repo, provider string, usedReferences map[string]bool) string {
-	reference := s.matchReferenceForNoteItem(release, itemText, owner, repo, provider)
-	if reference == "" || usedReferences[reference] {
-		return itemText
-	}
-
-	usedReferences[reference] = true
-	return fmt.Sprintf("%s (%s)", itemText, reference)
-}
-
-func (s *ReleaseService) collectReleaseReferences(release *models.Release, owner, repo, provider string, exclude map[string]bool) []string {
-	seen := make(map[string]bool)
-	refs := make([]string, 0)
-
-	appendItem := func(item models.ReleaseItem) {
-		ref := formatReleaseReference(item, owner, repo, provider)
-		if ref == "" || seen[ref] || exclude[ref] {
-			return
-		}
-		seen[ref] = true
-		refs = append(refs, ref)
-	}
-
-	for _, item := range release.Breaking {
-		appendItem(item)
-	}
-	for _, item := range release.Features {
-		appendItem(item)
-	}
-	for _, item := range release.BugFixes {
-		appendItem(item)
-	}
-	for _, item := range release.Improvements {
-		appendItem(item)
-	}
-	for _, item := range release.Documentation {
-		appendItem(item)
-	}
-	for _, item := range release.Other {
-		appendItem(item)
-	}
-
-	return refs
-}
-
-func (s *ReleaseService) matchReferenceForNoteItem(release *models.Release, itemText, owner, repo, provider string) string {
-	normalizedItem := normalizeChangelogMatchText(itemText)
-	if len(normalizedItem) < 12 {
-		return ""
-	}
-
-	for _, releaseItem := range s.allReleaseItems(release) {
-		normalizedDescription := normalizeChangelogMatchText(releaseItem.Description)
-		if len(normalizedDescription) < 12 {
-			continue
-		}
-
-		if normalizedItem == normalizedDescription || strings.Contains(normalizedItem, normalizedDescription) || strings.Contains(normalizedDescription, normalizedItem) {
-			return formatReleaseReference(releaseItem, owner, repo, provider)
-		}
-	}
-
-	return ""
-}
-
-func (s *ReleaseService) allReleaseItems(release *models.Release) []models.ReleaseItem {
-	items := make([]models.ReleaseItem, 0, len(release.Breaking)+len(release.Features)+len(release.BugFixes)+len(release.Improvements)+len(release.Documentation)+len(release.Other))
-	items = append(items, release.Breaking...)
-	items = append(items, release.Features...)
-	items = append(items, release.BugFixes...)
-	items = append(items, release.Improvements...)
-	items = append(items, release.Documentation...)
-	items = append(items, release.Other...)
-	return items
-}
-
-func normalizeChangelogMatchText(text string) string {
-	text = strings.ToLower(text)
-	replacer := strings.NewReplacer(
-		"`", " ",
-		"*", " ",
-		"_", " ",
-		"#", " ",
-		"-", " ",
-		"/", " ",
-		":", " ",
-		";", " ",
-		",", " ",
-		".", " ",
-		"(", " ",
-		")", " ",
-	)
-	text = replacer.Replace(text)
-	return strings.Join(strings.Fields(text), " ")
-}
-
-func formatReleaseReference(item models.ReleaseItem, owner, repo, provider string) string {
-	if item.PRNumber != "" {
-		if provider == "github" && owner != "" && repo != "" {
-			return fmt.Sprintf("[#%s](https://github.com/%s/%s/pull/%s)", item.PRNumber, owner, repo, item.PRNumber)
-		}
-		return fmt.Sprintf("#%s", item.PRNumber)
-	}
-
-	if item.CommitHash == "" {
-		return ""
-	}
-
-	shortHash := item.CommitHash
-	if len(shortHash) > 7 {
-		shortHash = shortHash[:7]
-	}
-
-	if provider == "github" && owner != "" && repo != "" {
-		return fmt.Sprintf("[`%s`](https://github.com/%s/%s/commit/%s)", shortHash, owner, repo, item.CommitHash)
-	}
-
-	return fmt.Sprintf("`%s`", shortHash)
-}
-
-// BuildChangelogPreview generates a preview of how the CHANGELOG entry will look
-func (s *ReleaseService) BuildChangelogPreview(ctx context.Context, release *models.Release, notes *models.ReleaseNotes) string {
-	return s.buildChangelogFromNotes(ctx, release, notes)
-}
-
-// buildChangelog formats the changelog from raw commits (fallback when AI is not available)
-func (s *ReleaseService) buildChangelog(release *models.Release) string {
-	var sb strings.Builder
-
-	sb.WriteString(fmt.Sprintf("## %s\n\n", release.Version))
-
-	if len(release.Breaking) > 0 {
-		sb.WriteString("### ⚠️ BREAKING CHANGES\n\n")
-		for _, item := range release.Breaking {
-			sb.WriteString(s.formatReleaseItem(item))
-		}
-		sb.WriteString("\n")
-	}
-
-	if len(release.Features) > 0 {
-		sb.WriteString("### ✨ New Features\n\n")
-		for _, item := range release.Features {
-			sb.WriteString(s.formatReleaseItem(item))
-		}
-		sb.WriteString("\n")
-	}
-
-	if len(release.BugFixes) > 0 {
-		sb.WriteString("### 🐛 Bug Fixes\n\n")
-		for _, item := range release.BugFixes {
-			sb.WriteString(s.formatReleaseItem(item))
-		}
-		sb.WriteString("\n")
-	}
-
-	if len(release.Improvements) > 0 {
-		sb.WriteString("### 🔧 Improvements\n\n")
-		for _, item := range release.Improvements {
-			sb.WriteString(s.formatReleaseItem(item))
-		}
-		sb.WriteString("\n")
-	}
-
-	if len(release.Documentation) > 0 {
-		sb.WriteString("### 📚 Documentation\n\n")
-		for _, item := range release.Documentation {
-			sb.WriteString(s.formatReleaseItem(item))
-		}
-		sb.WriteString("\n")
-	}
-
-	return sb.String()
-}
-
-func (s *ReleaseService) formatReleaseItem(item models.ReleaseItem) string {
-	line := "- "
-
-	if item.Scope != "" {
-		line += fmt.Sprintf("**%s**: ", item.Scope)
-	}
-
-	line += item.Description
-
-	if ref := formatReleaseReference(item, "", "", ""); ref != "" {
-		line += fmt.Sprintf(" (%s)", ref)
-	}
-
-	line += "\n"
-	return line
-}
-
-func (s *ReleaseService) CommitChangelog(ctx context.Context, version string) error {
-	log := logger.FromContext(ctx)
-	log.Info("starting changelog commit process", "version", version)
-
-	repoRoot, err := s.git.GetRepoRoot(ctx)
-	if err != nil {
-		log.Error("failed to get repository root for committing changelog", "error", err)
-		return domainErrors.NewAppError(domainErrors.TypeGit, "failed to get repository root", err)
-	}
-	changelogFile := filepath.Join(repoRoot, "CHANGELOG.md")
-
-	if err := s.git.AddFileToStaging(ctx, changelogFile); err != nil {
-		log.Error("failed to add CHANGELOG.md to staging", "error", err)
-		return domainErrors.NewAppError(domainErrors.TypeGit, "failed to add CHANGELOG.md to staging", err)
-	}
-	log.Debug("CHANGELOG.md added to staging")
-
-	versionFile, _, err := s.FindVersionFile(ctx)
-	if err != nil {
-		if s.config != nil && s.config.VersionFile != "" {
-			versionFile = s.config.VersionFile
-		} else {
-			versionFile = ""
-		}
-	}
-
-	if versionFile != "" {
-		if _, err := os.Stat(versionFile); err == nil {
-			log.Debug("adding version file to staging", "file", versionFile)
-			if err := s.git.AddFileToStaging(ctx, versionFile); err != nil {
-				log.Error("failed to add version file to staging", "file", versionFile, "error", err)
-				return domainErrors.NewAppError(domainErrors.TypeGit, fmt.Sprintf("failed to add version file to staging: %s", versionFile), err)
-			}
-			log.Debug("version file added to staging", "file", versionFile)
-		} else {
-			log.Debug("version file not found, skipping", "file", versionFile)
-		}
-	}
-
-	log.Debug("checking for staged changes")
-	if !s.git.HasStagedChanges(ctx) {
-		log.Error("no staged changes detected after adding files")
-		return domainErrors.ErrNoChanges
-	}
-	log.Debug("staged changes detected, proceeding with commit")
-
-	message := fmt.Sprintf("chore: update changelog and bump version to %s", version)
-	if err := s.git.CreateCommit(ctx, message); err != nil {
-		log.Error("failed to create commit", "error", err)
-		return domainErrors.NewAppError(domainErrors.TypeGit, "failed to commit changelog and version bump", err)
-	}
-
-	log.Info("changelog commit process completed successfully")
-	return nil
-
-}
-
-// PushChanges pushes committed changes to the remote repository
-func (s *ReleaseService) PushChanges(ctx context.Context) error {
-	return s.git.Push(ctx)
-}
-
-func (s *ReleaseService) UpdateAppVersion(ctx context.Context, version string) error {
-	log := logger.FromContext(ctx)
-
-	versionFile, versionPattern, err := s.FindVersionFile(ctx)
-	if err != nil {
-		if s.config == nil || s.config.VersionFile == "" {
-			log.Warn("could not auto-detect version file, skipping version update", "error", err)
-			return nil
-		}
-
-		log.Warn("could not auto-detect version file, using configured version file", "error", err)
-		versionFile = s.config.VersionFile
-
-		if s.config != nil {
-			if s.config.VersionPattern != "" {
-				versionPattern = s.config.VersionPattern
-			}
-		}
-	}
-
-	log.Debug("updating version",
-		"file", versionFile,
-		"pattern", versionPattern,
-		"new_version", version)
-
-	content, err := os.ReadFile(versionFile)
-	if err != nil {
-		if os.IsNotExist(err) && (s.config == nil || s.config.VersionFile == "") {
-			log.Warn("version file not found, skipping version update", "file", versionFile)
-			return nil
-		}
-		return domainErrors.NewAppError(domainErrors.TypeInternal,
-			fmt.Sprintf("failed to read version file: %s", versionFile), err)
-	}
-
-	re, err := regexp.Compile(versionPattern)
-	if err != nil {
-		return domainErrors.NewAppError(domainErrors.TypeInternal,
-			fmt.Sprintf("invalid version pattern: %s", versionPattern), err)
-	}
-
-	currentContent := string(content)
-	if !re.MatchString(currentContent) {
-		return domainErrors.NewAppError(domainErrors.TypeInternal,
-			fmt.Sprintf("version pattern not found in %s with pattern %s", versionFile, versionPattern), nil)
-	}
-
-	match := re.FindString(currentContent)
-	if match == "" {
-		return domainErrors.NewAppError(domainErrors.TypeInternal,
-			"could not find version match", nil)
-	}
-
-	cleanVersion := strings.TrimPrefix(version, "v")
-
-	var newMatch string
-	ext := filepath.Ext(versionFile)
-
-	switch ext {
-	case ".json":
-		versionRe := regexp.MustCompile(`"version"\s*:\s*"([^"]+)"`)
-		newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`"version": "%s"`, cleanVersion))
-	case ".toml":
-		versionRe := regexp.MustCompile(`version\s*=\s*"([^"]+)"`)
-		newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`version = "%s"`, cleanVersion))
-	case ".xml":
-		versionRe := regexp.MustCompile(`<version>([^<]+)</version>`)
-		newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`<version>%s</version>`, cleanVersion))
-	case ".py":
-		if strings.Contains(match, `"`) {
-			versionRe := regexp.MustCompile(`"([^"]+)"`)
-			newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`"%s"`, cleanVersion))
-		} else if strings.Contains(match, `'`) {
-			versionRe := regexp.MustCompile(`'([^']+)'`)
-			newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`'%s'`, cleanVersion))
-		}
-	case ".go", ".js", ".ts", ".rs", ".php", ".rb":
-		if strings.Contains(match, `"`) {
-			valMatch := regex.QuotedString.FindStringIndex(match)
-			if valMatch != nil {
-				newMatch = match[:valMatch[0]] + fmt.Sprintf(`"%s"`, cleanVersion) + match[valMatch[1]:]
-			} else {
-				versionRe := regexp.MustCompile(`"([^"]+)"`)
-				newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`"%s"`, cleanVersion))
-			}
-		} else if strings.Contains(match, `'`) {
-			versionRe := regexp.MustCompile(`'([^']+)'`)
-			newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`'%s'`, cleanVersion))
-		} else {
-			versionRe := regexp.MustCompile(`[\d.]+`)
-			newMatch = versionRe.ReplaceAllString(match, cleanVersion)
-		}
-	default:
-		if strings.Contains(match, `"`) {
-			versionRe := regexp.MustCompile(`"([^"]+)"`)
-			newMatch = versionRe.ReplaceAllString(match, fmt.Sprintf(`"%s"`, cleanVersion))
-		} else {
-			versionRe := regexp.MustCompile(`[\d.]+`)
-			newMatch = versionRe.ReplaceAllString(match, cleanVersion)
-		}
-	}
-
-	if newMatch == "" {
-		newMatch = match
-	}
-
-	newContent := strings.Replace(currentContent, match, newMatch, 1)
-
-	if err := os.WriteFile(versionFile, []byte(newContent), 0644); err != nil {
-		return domainErrors.NewAppError(domainErrors.TypeInternal,
-			fmt.Sprintf("failed to write version file: %s", versionFile), err)
-	}
-
-	log.Info("version updated successfully",
-		"file", versionFile,
-		"version", version)
-
-	return nil
-}
-
-func (s *ReleaseService) FindVersionFile(ctx context.Context) (string, string, error) {
-	log := logger.FromContext(ctx)
-
-	if s.versionFileCache.file != "" && s.versionFileCache.pattern != "" {
-		log.Debug("using cached version file",
-			"file", s.versionFileCache.file,
-			"pattern", s.versionFileCache.pattern,
-		)
-		return s.versionFileCache.file, s.versionFileCache.pattern, nil
-	}
-
-	if s.config != nil && s.config.VersionFile != "" {
-		pattern := s.config.VersionPattern
-		if pattern == "" {
-			detectedPattern, err := s.detectPatternInFile(s.config.VersionFile)
-			if err == nil && detectedPattern != "" {
-				pattern = detectedPattern
-			} else {
-				pattern = `Version:\s*".*"`
-			}
-		}
-		log.Debug("using configured version file",
-			"file", s.config.VersionFile,
-			"pattern", pattern)
-
-		s.versionFileCache.file = s.config.VersionFile
-		s.versionFileCache.pattern = pattern
-
-		return s.config.VersionFile, pattern, nil
-	}
-
-	projectType := s.detectProjectType()
-	log.Debug("detected project type", "type", projectType)
-
-	if files, ok := versionFilesByLanguage[projectType]; ok {
-		for _, filePath := range files {
-			if strings.Contains(filePath, "*") {
-				matches, err := filepath.Glob(filePath)
-				if err == nil && len(matches) > 0 {
-					filePath = matches[0]
-				} else {
-					continue
-				}
-			}
-
-			if _, err := os.Stat(filePath); err == nil {
-				pattern, err := s.detectPatternInFileForLanguage(filePath, projectType)
-				if err == nil && pattern != "" {
-					log.Debug("version file found automatically",
-						"file", filePath,
-						"pattern", pattern,
-						"language", projectType,
-					)
-
-					s.versionFileCache.file = filePath
-					s.versionFileCache.pattern = pattern
-					s.versionFileCache.lang = projectType
-					return filePath, pattern, nil
-				}
-			}
-		}
-	}
-
-	foundFile, foundPattern, err := s.searchVersionFileRecursive(projectType)
-	if err == nil && foundFile != "" {
-		log.Debug("version file found recursively",
-			"file", foundFile,
-			"pattern", foundPattern,
-		)
-
-		s.versionFileCache.file = foundFile
-		s.versionFileCache.pattern = foundPattern
-		s.versionFileCache.lang = projectType
-
-		return foundFile, foundPattern, nil
-	}
-
-	if files, ok := versionFilesByLanguage[projectType]; ok {
-		return "", "", domainErrors.ErrVersionFileNotFound.
-			WithContext("project_type", projectType).
-			WithContext("searched_paths", files)
-	}
-
-	return "", "", domainErrors.ErrVersionFileNotFound.
-		WithContext("project_type", projectType)
-}
-
 func (s *ReleaseService) ValidateMainBranch(ctx context.Context) error {
 	log := logger.FromContext(ctx)
 
@@ -1443,498 +330,4 @@ func (s *ReleaseService) ValidateMainBranch(ctx context.Context) error {
 		"branch", branch,
 	)
 	return nil
-}
-
-func (s *ReleaseService) filterValidCommits(commits []models.Commit) []models.Commit {
-	var valid []models.Commit
-	for _, commit := range commits {
-		if regex.ConventionalCommit.MatchString(commit.Message) {
-			valid = append(valid, commit)
-		}
-	}
-	return valid
-}
-
-func (s *ReleaseService) validateVersionIncrement(oldVersion, newVersion string) error {
-	oldClean := strings.TrimPrefix(oldVersion, "v")
-	newClean := strings.TrimPrefix(newVersion, "v")
-
-	if semver.Compare("v"+oldClean, "v"+newClean) >= 0 {
-		return fmt.Errorf("new version %s must be greater than previous version %s", newVersion, oldVersion)
-	}
-	return nil
-}
-
-func (s *ReleaseService) validateVersionString(versionStr string) error {
-	cleanVersion := strings.TrimPrefix(versionStr, "v")
-
-	if !semver.IsValid("v" + cleanVersion) {
-		return fmt.Errorf("invalid semver format: %s", versionStr)
-	}
-	return nil
-}
-
-type versionPatternInfo struct {
-	name               string
-	detectionPattern   string
-	replacementPattern string
-}
-
-var consolidatedVersionPatterns = map[string][]versionPatternInfo{
-	"go": {
-		{
-			name:               "const Version",
-			detectionPattern:   `const\s+Version\s*=\s*"([^"]+)"`,
-			replacementPattern: `const\s+Version\s*=\s*"[^"]*"`,
-		},
-		{
-			name:               "var Version",
-			detectionPattern:   `var\s+Version\s*=\s*"([^"]+)"`,
-			replacementPattern: `var\s+Version\s*=\s*"[^"]*"`,
-		},
-		{
-			name:               "const Version unquoted",
-			detectionPattern:   `const\s+Version\s*=\s*([\d.]+)`,
-			replacementPattern: `const\s+Version\s*=\s*[\d.]+`,
-		},
-		{
-			name:               "var Version unquoted",
-			detectionPattern:   `var\s+Version\s*=\s*([\d.]+)`,
-			replacementPattern: `var\s+Version\s*=\s*[\d.]+`,
-		},
-		{
-			name:               "Version:",
-			detectionPattern:   `Version:\s*"([^"]+)"`,
-			replacementPattern: `Version:\s*"[^"]*"`,
-		},
-		{
-			name:               "Version =",
-			detectionPattern:   `Version\s*=\s*"([^"]+)"`,
-			replacementPattern: `Version\s*=\s*"[^"]*"`,
-		},
-	},
-	"python": {
-		{
-			name:               "__version__ double quotes",
-			detectionPattern:   `__version__\s*=\s*"([^"]+)"`,
-			replacementPattern: `__version__\s*=\s*"[^"]*"`,
-		},
-		{
-			name:               "__version__ single quotes",
-			detectionPattern:   `__version__\s*=\s*'([^']+)'`,
-			replacementPattern: `__version__\s*=\s*'[^']*'`,
-		},
-		{
-			name:               "version double quotes",
-			detectionPattern:   `version\s*=\s*"([^"]+)"`,
-			replacementPattern: `version\s*=\s*"[^"]*"`,
-		},
-		{
-			name:               "version any quotes",
-			detectionPattern:   `version\s*=\s*['"]([^'"]+)['"]`,
-			replacementPattern: `version\s*=\s*['"][^'"]*['"]`,
-		},
-		{
-			name:               "VERSION",
-			detectionPattern:   `VERSION\s*=\s*"([^"]+)"`,
-			replacementPattern: `VERSION\s*=\s*"[^"]*"`,
-		},
-	},
-	"js": {
-		{
-			name:               "version in JSON",
-			detectionPattern:   `"version"\s*:\s*"([^"]+)"`,
-			replacementPattern: `"version"\s*:\s*"[^"]*"`,
-		},
-		{
-			name:               "version in JSON single quotes",
-			detectionPattern:   `'version'\s*:\s*'([^']+)'`,
-			replacementPattern: `'version'\s*:\s*'[^']*'`,
-		},
-		{
-			name:               "export const version",
-			detectionPattern:   `export\s+const\s+version\s*=\s*"([^"]+)"`,
-			replacementPattern: `export\s+const\s+version\s*=\s*"[^"]*"`,
-		},
-		{
-			name:               "export const version single quotes",
-			detectionPattern:   `export\s+const\s+version\s*=\s*'([^']+)'`,
-			replacementPattern: `export\s+const\s+version\s*=\s*'[^']*'`,
-		},
-	},
-	"rust": {
-		{
-			name:               "version in TOML",
-			detectionPattern:   `version\s*=\s*"([^"]+)"`,
-			replacementPattern: `version\s*=\s*"[^"]*"`,
-		},
-	},
-	"java": {
-		{
-			name:               "version in XML",
-			detectionPattern:   `<version>([^<]+)</version>`,
-			replacementPattern: `<version>[^<]+</version>`,
-		},
-		{
-			name:               "version in properties",
-			detectionPattern:   `version\s*=\s*['"]([^'"]+)['"]`,
-			replacementPattern: `version\s*=\s*['"][^'"]*['"]`,
-		},
-	},
-	"csharp": {
-		{
-			name:               "AssemblyVersion",
-			detectionPattern:   `AssemblyVersion\s*\(\s*"([^"]+)"`,
-			replacementPattern: `AssemblyVersion\s*\(\s*"[^"]*"`,
-		},
-		{
-			name:               "Version in XML",
-			detectionPattern:   `<Version>([^<]+)</Version>`,
-			replacementPattern: `<Version>[^<]+</Version>`,
-		},
-		{
-			name:               "version in JSON",
-			detectionPattern:   `"version"\s*:\s*"([^"]+)"`,
-			replacementPattern: `"version"\s*:\s*"[^"]*"`,
-		},
-	},
-	"php": {
-		{
-			name:               "version in JSON",
-			detectionPattern:   `"version"\s*:\s*"([^"]+)"`,
-			replacementPattern: `"version"\s*:\s*"[^"]*"`,
-		},
-		{
-			name:               "const VERSION",
-			detectionPattern:   `const\s+VERSION\s*=\s*['"]([^'"]+)['"]`,
-			replacementPattern: `const\s+VERSION\s*=\s*['"][^'"]*['"]`,
-		},
-	},
-	"ruby": {
-		{
-			name:               "VERSION",
-			detectionPattern:   `VERSION\s*=\s*['"]([^'"]+)['"]`,
-			replacementPattern: `VERSION\s*=\s*['"][^'"]*['"]`,
-		},
-		{
-			name:               ".version",
-			detectionPattern:   `\.version\s*=\s*['"]([^'"]+)['"]`,
-			replacementPattern: `\.version\s*=\s*['"][^'"]*['"]`,
-		},
-	},
-}
-
-func (s *ReleaseService) detectPatternInFile(filePath string) (string, error) {
-	content, err := os.ReadFile(filePath)
-	if err != nil {
-		return "", err
-	}
-
-	fileContent := string(content)
-	lang := detectLanguageFromFile(filePath)
-
-	if patterns, ok := consolidatedVersionPatterns[lang]; ok {
-		for _, patternInfo := range patterns {
-			re, err := regexp.Compile(patternInfo.detectionPattern)
-			if err != nil {
-				continue
-			}
-			if re.MatchString(fileContent) {
-				matches := re.FindStringSubmatch(fileContent)
-				if len(matches) > 1 {
-					if err := s.validateVersionString(matches[1]); err != nil {
-						continue
-					}
-				}
-				return patternInfo.replacementPattern, nil
-			}
-		}
-	}
-
-	for _, patterns := range consolidatedVersionPatterns {
-		for _, patternInfo := range patterns {
-			re, err := regexp.Compile(patternInfo.detectionPattern)
-			if err != nil {
-				continue
-			}
-			if re.MatchString(fileContent) {
-				matches := re.FindStringSubmatch(fileContent)
-				if len(matches) > 1 {
-					if err := s.validateVersionString(matches[1]); err != nil {
-						continue
-					}
-				}
-				return patternInfo.replacementPattern, nil
-			}
-		}
-	}
-
-	return "", fmt.Errorf("no version pattern found in file")
-}
-
-func (s *ReleaseService) searchVersionFileRecursive(lang string) (string, string, error) {
-	searchDirs := map[string][]string{
-		"go":     {"internal", "pkg", "version", "cmd"},
-		"python": {"src", "lib", "."},
-		"js":     {"src", "lib", "."},
-		"rust":   {"."},
-		"java":   {"src", "."},
-		"csharp": {"Properties", "."},
-		"php":    {"src", "."},
-		"ruby":   {"lib", "."},
-	}
-
-	dirs, ok := searchDirs[lang]
-	if !ok {
-		dirs = []string{"."}
-	}
-
-	extensions := map[string][]string{
-		"go":     {".go"},
-		"python": {".py"},
-		"js":     {".js", ".ts"},
-		"rust":   {".rs", ".toml"},
-		"java":   {".xml", ".gradle", ".properties"},
-		"csharp": {".cs", ".csproj", ".props"},
-		"php":    {".php", ".json"},
-		"ruby":   {".rb", ".gemspec"},
-	}
-
-	exts, ok := extensions[lang]
-	if !ok {
-		exts = []string{""}
-	}
-
-	filesScanned := 0
-
-	for _, dir := range dirs {
-		var foundPath string
-		var foundPattern string
-
-		err := filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				return nil
-			}
-
-			if filesScanned >= maxFilesToScan {
-				return filepath.SkipDir
-			}
-
-			if info.IsDir() {
-				if ignoreDirs[info.Name()] {
-					return filepath.SkipDir
-				}
-
-				depth := len(strings.Split(path, string(os.PathSeparator)))
-				if depth > maxSearchDepth {
-					return filepath.SkipDir
-				}
-				return nil
-			}
-
-			filesScanned++
-
-			hasValidExt := false
-			for _, ext := range exts {
-				if ext == "" || strings.HasSuffix(path, ext) {
-					hasValidExt = true
-					break
-				}
-			}
-			if !hasValidExt {
-				return nil
-			}
-
-			if strings.Contains(strings.ToLower(path), "version") ||
-				strings.Contains(strings.ToLower(info.Name()), "version") ||
-				path == "package.json" || path == "Cargo.toml" || path == "setup.py" {
-				pattern, err := s.detectPatternInFileForLanguage(path, lang)
-				if err == nil && pattern != "" {
-					foundPath = path
-					foundPattern = pattern
-					return fmt.Errorf("found: %s", path)
-				}
-			}
-			return nil
-		})
-
-		if err != nil && strings.HasPrefix(err.Error(), "found: ") {
-			return foundPath, foundPattern, nil
-		}
-	}
-
-	return "", "", fmt.Errorf("version file not found")
-}
-
-func (s *ReleaseService) detectProjectType() string {
-	indicators := map[string][]string{
-		"go":     {"go.mod", "Gopkg.toml", "glide.yaml"},
-		"python": {"setup.py", "pyproject.toml", "requirements.txt", "Pipfile"},
-		"js":     {"package.json", "yarn.lock", "package-lock.json"},
-		"rust":   {"Cargo.toml"},
-		"java":   {"pom.xml", "build.gradle", "build.gradle.kts"},
-		"csharp": {".csproj", ".sln", "project.json"},
-		"php":    {"composer.json"},
-		"ruby":   {"Gemfile", "Rakefile"},
-	}
-
-	for lang, files := range indicators {
-		for _, file := range files {
-			if _, err := os.Stat(file); err == nil {
-				return lang
-			}
-		}
-	}
-
-	if hasFilesWithExtension(".go") {
-		return "go"
-	}
-	if hasFilesWithExtension(".py") {
-		return "python"
-	}
-	if hasFilesWithExtension(".js") || hasFilesWithExtension(".ts") {
-		return "js"
-	}
-	if hasFilesWithExtension(".rs") {
-		return "rust"
-	}
-
-	return "unknown"
-}
-
-func hasFilesWithExtension(ext string) bool {
-	entries, err := os.ReadDir(".")
-	if err != nil {
-		return false
-	}
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ext) {
-			return true
-		}
-	}
-	return false
-}
-
-var versionFilesByLanguage = map[string][]string{
-	"go": {
-		"internal/version/version.go",
-		"pkg/version/version.go",
-		"version/version.go",
-		"version.go",
-		"cmd/main.go",
-		"main.go",
-		"internal/version.go",
-		"pkg/version.go",
-	},
-	"python": {
-		"__version__.py",
-		"version.py",
-		"src/*/__version__.py",
-		"setup.py",
-		"pyproject.toml",
-	},
-	"js": {
-		"package.json",
-		"src/version.js",
-		"src/version.ts",
-		"lib/version.js",
-	},
-	"rust": {
-		"Cargo.toml",
-	},
-	"java": {
-		"pom.xml",
-		"build.gradle",
-		"build.gradle.kts",
-		"src/main/resources/version.properties",
-	},
-	"csharp": {
-		"Properties/AssemblyInfo.cs",
-		"Directory.Build.props",
-		"*.csproj",
-	},
-	"php": {
-		"composer.json",
-		"src/Version.php",
-	},
-	"ruby": {
-		"lib/*/version.rb",
-		"version.rb",
-		"*.gemspec",
-	},
-}
-
-func (s *ReleaseService) detectPatternInFileForLanguage(filePath, lang string) (string, error) {
-	content, err := os.ReadFile(filePath)
-	if err != nil {
-		return "", err
-	}
-
-	fileContent := string(content)
-	patterns, ok := consolidatedVersionPatterns[lang]
-	if !ok {
-		return s.detectPatternInFile(filePath)
-	}
-
-	for _, patternInfo := range patterns {
-		re, err := regexp.Compile(patternInfo.detectionPattern)
-		if err != nil {
-			continue
-		}
-		if re.MatchString(fileContent) {
-			matches := re.FindStringSubmatch(fileContent)
-			if len(matches) > 1 {
-				versionValue := matches[1]
-				if err := s.validateVersionString(versionValue); err != nil {
-					continue
-				}
-			}
-			return patternInfo.replacementPattern, nil
-		}
-	}
-
-	return "", fmt.Errorf("no version pattern found in file")
-}
-
-func detectLanguageFromFile(filePath string) string {
-	ext := filepath.Ext(filePath)
-	extToLang := map[string]string{
-		".go":      "go",
-		".py":      "python",
-		".js":      "js",
-		".ts":      "js",
-		".rs":      "rust",
-		".toml":    "rust",
-		".xml":     "java",
-		".cs":      "csharp",
-		".csproj":  "csharp",
-		".props":   "csharp",
-		".php":     "php",
-		".rb":      "ruby",
-		".gemspec": "ruby",
-		".json":    "js",
-	}
-
-	filename := strings.ToLower(filepath.Base(filePath))
-	if filename == "package.json" || filename == "package-lock.json" {
-		return "js"
-	}
-	if filename == "composer.json" {
-		return "php"
-	}
-	if filename == "cargo.toml" {
-		return "rust"
-	}
-	if filename == "pom.xml" {
-		return "java"
-	}
-	if filename == "setup.py" || filename == "pyproject.toml" {
-		return "python"
-	}
-
-	if lang, ok := extToLang[ext]; ok {
-		return lang
-	}
-
-	return "unknown"
 }
